requirements.txt
streamlit
langchain_openai
youtube_transcript_api
re
requests
python-dotenv
os
langchain_core